renamed api to app and udpated locations to what will be the next project paths
