+++ CarInventoryApi/src/App/ErrorHandlers/ValidationExceptionHandler.cs
@@ -2,29 +2,30 @@
 using Microsoft.AspNetCore.Diagnostics;
 using Microsoft.AspNetCore.Mvc;
 
-namespace CarInventory.Api.ErrorHandlers;
+namespace CarInventory.App.ErrorHandlers;
 
-public sealed class BadHttpRequestExceptionHandler : IExceptionHandler
+public class ValidationExceptionHandler : IExceptionHandler
 {
     /// <inheritdoc />
     public async ValueTask<bool> TryHandleAsync(HttpContext httpContext, Exception exception, CancellationToken cancellationToken)
     {
-        if (exception is not BadHttpRequestException)
+        if (exception is not FluentValidation.ValidationException)
         {
             return false;
         }
 
-        httpContext.Response.StatusCode = (int)HttpStatusCode.BadRequest;
+        httpContext.Response.StatusCode = (int)HttpStatusCode.RequestTimeout;
 
         await httpContext.Response.WriteAsJsonAsync(new ProblemDetails
         {
             Status = (int)HttpStatusCode.BadRequest,
             Type = exception.GetType().Name,
             Title = "Invalid Request",
-            Detail = "One or more problems occurred with the request, please check parameters and try again",
+            Detail = exception.Message,
             Instance = $"{httpContext.Request.Method} {httpContext.Request.Path}"
 
         }, cancellationToken);
         return true;
+
     }
 }
